Refactor: Mj_models -> prx_models: From now on, it will handle both sim and real              robot-specific files
New structure according to last meeting:
  * Infrastructure (aka interface)
    * prx_models
    * ML4KP_bridge
  * Task Planning
  * Motion Planning
  * World
    * Mujoco-ros

Adding READMEs to the four directories
