+++ infrastructure/README.md
@@ -1,0 +1,11 @@
+# Infrastructure
+(Or interface if needed at some point)
+
+## PRX Models
+	* Robot Models, either real or for simulation
+ 	* Robot-specific messages/services
+
+## ML4KP Bridge
+	* Import of ML4KP
+	
+
