+++ infrastructure/prx_models/models/friction_floors/friction_floor_4x4.xml
