+++ infrastructure/prx_models/models/mushr/meshes/mushr_wheel.stl
