+++ infrastructure/prx_models/models/mushr/meshes/mushr_ydlidar.stl
