+++ infrastructure/prx_models/msg/mushr/MushrPlan.msg
@@ -1,3 +1,3 @@
 std_msgs/Header header
-mj_models/MushrControl[] controls
+prx_models/MushrControl[] controls
 std_msgs/Float64[] durations
