+++ world/README.md
@@ -1,0 +1,8 @@
+# World
+Packages related to real robots or simulation
+
+## Mujoco Ros
+	* Mujoco execution with ROS
+
+## Mushr 
+ToDo
