+++ world/mujoco_ros/launch/mushr_open_loop.launch
