+++ world/mujoco_ros/package.xml
@@ -50,17 +50,17 @@
   <!--   <doc_depend>doxygen</doc_depend> -->
   <buildtool_depend>catkin</buildtool_depend>
   
+  <build_depend>message_generation</build_depend>
+  <build_depend>prx_models</build_depend>
   <build_depend>roscpp</build_depend>
   <build_depend>std_msgs</build_depend>
-  <build_depend>message_generation</build_depend>
-  <build_depend>mj_models</build_depend>
   
-  <build_export_depend>mj_models</build_export_depend>
+  <build_export_depend>prx_models</build_export_depend>
   <build_export_depend>roscpp</build_export_depend>
   <build_export_depend>std_msgs</build_export_depend>
   
   <exec_depend>message_runtime</exec_depend>
-  <exec_depend>mj_models</exec_depend>
+  <exec_depend>prx_models</exec_depend>
   <exec_depend>roscpp</exec_depend>
   <exec_depend>std_msgs</exec_depend>
