+++ world/mujoco_ros/scripts/mushr_open_loop_publisher.py
